Rewrite/refactor and increase frequencey of running mnowatch
mnowatch will now run hourly on the final day of voting and will run immediately after the voting closes.  Lots of refactoring in the rewrite should be easier to extend it in the future if required.  Added additional sanity checks to abort this program if the environment is not correct.